xsm: switch XSM init to boot info structures
Change the XSM implementation to use the boot info structure instead of
the multiboot module_map.

Drops a dependency on CONFIG_MULTIBOOT, so boot module logic is now
used whenever the DEVICE_TREE specific logic (for Arm) is not, with the
bootinfo header conditionally included to ensure no change on Arm.

Adds a multiboot header inclusion into guest/xen/pvh-boot.c
since it is no longer provided via transitive inclusion and the source
in that file uses multiboot structures.

Signed-off-by: Christopher Clark <christopher.w.clark@gmail.com>
Signed-off-by: Daniel P. Smith <dpsmith@apertussolutions.com>