x86, arm, riscv: add per-arch bootinfo headers
Changes to ensure that inclusion of <xen/bootinfo.h> succeeds in each
of the main archtecture builds and in the 32-bit early x86 boot build.

The <xen/bootinfo.h> header contains structures that will be used in efi
logic prior to the main start of Xen, so it needs to be suitable for
inclusion in Arm source files.

The same header will also be included in early x86 boot logic, where
asm/* headers are not reachable, and so they cannot be transitively
included, so a ifdef condition is needed to prevent an attempt to
include <asm/bootinfo.h> from within <xen/bootinfo.h> in that case.

The early x86 boot logic can then directly include <asm/bootinfo.h> via
a specified directory path where it can be included successfully.

Signed-off-by: Christopher Clark <christopher.w.clark@gmail.com>
Signed-off-by: Daniel P. Smith <dpsmith@apertussolutions.com>