docs: update hyperlaunch device tree
With on going development of hyperlaunch, changes to the device tree definitions
has been necessary. This commit updates the specification for all current changes
along with changes expected to be made in finalizing the capability.

This commit also adds a HYPERLAUNCH section to the MAINTAINERS file and places
this documentation under its purview. It also reserves the path
`xen/common/domain-builder` for the hyperlaunch domain builder code base.

Signed-off-by: Daniel P. Smith <dpsmith@apertussolutions.com>